Extract common functionality to one class
