+++ brew/src/main/java/nl/tudelft/serg/evosql/brew/generator/junit/JUnit5TestGenerator.java
@@ -13,252 +13,18 @@ import javax.lang.model.element.Modifier;
 import java.util.HashSet;
 import java.util.Set;
 
-public class JUnit5TestGenerator implements Generator {
+public class JUnit5TestGenerator extends JUnitGenerator {
 
-    private final JUnitGeneratorSettings jUnitGeneratorSettings;
-
-    /**
-     * Initializes a new instance of the JUnit4TestGenerator class.
-     *
-     * @param junitgeneratorsettings The settings used for this generator.
-     */
-    public JUnit5TestGenerator(JUnitGeneratorSettings junitgeneratorsettings) {
-        this.jUnitGeneratorSettings = junitgeneratorsettings;
-    }
-
-    @Override
-    public String generate(Result result, VendorOptions vendorOptions) {
-        AnnotationSpec generatedAnnotation = AnnotationSpec.builder(Generated.class)
-                .addMember("value", "$S", getClass().getCanonicalName())
-                .build();
-        TypeSpec.Builder typeSpecBuilder = TypeSpec.classBuilder("EvoSQLQueryTest")
-                .addModifiers(Modifier.PUBLIC)
-                .addAnnotation(generatedAnnotation);
-
-        typeSpecBuilder.addMethod(generateRunSQL());
-        typeSpecBuilder.addMethod(generateCreateTables(result, vendorOptions));
-        typeSpecBuilder.addMethod(generateCleanTables(result, vendorOptions));
-        typeSpecBuilder.addMethod(generateDropTables(result, vendorOptions));
-
-        typeSpecBuilder.addMethod(generateBeforeAll());
-        typeSpecBuilder.addMethod(generateBeforeEach());
-        typeSpecBuilder.addMethod(generateAfterEach());
-        typeSpecBuilder.addMethod(generateAfterAll());
-
-        for (Path path : result.getPaths()) {
-            typeSpecBuilder.addMethod(generatePathTest(path, vendorOptions));
-        }
-
-        JavaFile javaFile = JavaFile.builder(jUnitGeneratorSettings.getFilePackage(), typeSpecBuilder.build()).build();
-        return javaFile.toString();
-    }
-
-    /**
-     * Generates a method specification for the runSQL method.
-     *
-     * @return method.
-     */
-    private MethodSpec generateRunSQL() {
-        return MethodSpec.methodBuilder("runSQL")
-                .addModifiers(Modifier.PRIVATE, Modifier.STATIC)
-                .returns(TypeName.BOOLEAN)
-                .addParameter(String.class, "query")
-                .addJavadoc(""
-                        + "This method should connect to your database and execute the given query.\n"
-                        + "In order for the assertions to work this method must return true in the case \n"
-                        + "that the query yields at least one result and false if there is no result.\n\n"
-                        + "@param  query sql query to execute\n"
-                        + "@return result of query.\n")
-                .addComment("TODO: Implement method stub")
-                .addStatement("return false")
-                .build();
-    }
-
-
-    /**
-     * Generates a method specification for a method that creates the tables required for queries.
-     *
-     * @param result        Result containing data.
-     * @param vendorOptions Vendor options for the database.
-     * @return A method specification for a method that drops the tables.
-     */
-    private MethodSpec generateCreateTables(Result result, VendorOptions vendorOptions) {
-        // Method signature
-        MethodSpec.Builder createTables = MethodSpec.methodBuilder("createTables")
-                .addModifiers(Modifier.PRIVATE, Modifier.STATIC)
-                .returns(TypeName.VOID)
-                .addJavadoc("Creates tables required for queries.\n");
-
-        // Create tables code
-        TableCreationBuilder tableCreationBuilder = new TableCreationBuilder(vendorOptions);
-        Set<String> tableCreateStrings = new HashSet<>();
-        result.getPaths().stream().map(tableCreationBuilder::buildQueries).forEach(tableCreateStrings::addAll);
-        for (String s : tableCreateStrings) {
-            createTables.addStatement("runSQL($S)", s);
-        }
-
-        return createTables.build();
-    }
-
-
-    /**
-     * Generates a method specification for a method that truncates the tables.
-     *
-     * @param result        Result containing data.
-     * @param vendorOptions Vendor options for the database
-     * @return A method specification for a method that drops the tables.
-     */
-    private MethodSpec generateCleanTables(Result result, VendorOptions vendorOptions) {
-        // Method signature
-        MethodSpec.Builder cleanTables = MethodSpec.methodBuilder("cleanTables")
-                .addModifiers(Modifier.PRIVATE, Modifier.STATIC)
-                .returns(TypeName.VOID)
-                .addJavadoc("Truncates the tables.\n");
-
-        // Create tables code
-        CleaningBuilder cleaningBuilder = new CleaningBuilder(vendorOptions);
-        Set<String> tableCleanStrings = new HashSet<>();
-        result.getPaths().stream().map(cleaningBuilder::buildQueries).forEach(tableCleanStrings::addAll);
-        for (String s : tableCleanStrings) {
-            cleanTables.addStatement("runSQL($S)", s);
-        }
-
-        return cleanTables.build();
-    }
-
-    /**
-     * Generates a method specification for a method that drops the tables.
-     *
-     * @param result        Result containing data.
-     * @param vendorOptions Vendor options for the database.
-     * @return A method specification for a method that drops the tables.
-     */
-    private MethodSpec generateDropTables(Result result, VendorOptions vendorOptions) {
-        // Method signature
-        MethodSpec.Builder dropTables = MethodSpec.methodBuilder("dropTables")
-                .addModifiers(Modifier.PRIVATE, Modifier.STATIC)
-                .returns(TypeName.VOID)
-                .addJavadoc("Drops the tables.\n");
-
-        // Create tables code
-        DestructionBuilder destructionBuilder = new DestructionBuilder(vendorOptions);
-        Set<String> destructionStrings = new HashSet<>();
-        result.getPaths().stream().map(destructionBuilder::buildQueries).forEach(destructionStrings::addAll);
-        for (String s : destructionStrings) {
-            dropTables.addStatement("runSQL($S)", s);
-        }
-
-        return dropTables.build();
-    }
-
-
-    /**
-     * Generates a method specification for a before all test method.
-     *
-     * @return A method specification for a before all test method.
-     */
-    private MethodSpec generateBeforeAll() {
-        MethodSpec.Builder beforeAll = MethodSpec.methodBuilder("beforeAll")
-                .addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-                .returns(TypeName.VOID)
-                .addAnnotation(BeforeAll.class);
-
-        if (jUnitGeneratorSettings.isCreateTablesBeforeRunning()) {
-            beforeAll.addStatement("createTables()");
-        }
-        return beforeAll.build();
-    }
-
-    /**
-     * Generates a method specification for a before each test method.
-     *
-     * @return A method specification for a before each test method.
-     */
-    private MethodSpec generateBeforeEach() {
-        MethodSpec.Builder beforeEach = MethodSpec.methodBuilder("beforeEach")
-                .addModifiers(Modifier.PUBLIC)
-                .returns(TypeName.VOID)
-                .addAnnotation(BeforeEach.class);
-
-        if (jUnitGeneratorSettings.isCleanTablesBeforeEachRun()) {
-            beforeEach.addStatement("cleanTables()");
-        }
-        return beforeEach.build();
-    }
-
-    /**
-     * Generates a method specification for an after each test method.
-     *
-     * @return A method specification for an after each test method.
-     */
-    private MethodSpec generateAfterEach() {
-        MethodSpec.Builder beforeEach = MethodSpec.methodBuilder("afterEach")
-                .addModifiers(Modifier.PUBLIC)
-                .returns(TypeName.VOID)
-                .addAnnotation(AfterEach.class);
-
-        if (jUnitGeneratorSettings.isCreateTablesBeforeRunning()) {
-            beforeEach.addStatement("cleanTables()");
-        }
-        return beforeEach.build();
-    }
-
-    /**
-     * Generates a method specification for an after all test method.
-     *
-     * @return A method specification for an after all test method.
-     */
-    private MethodSpec generateAfterAll() {
-        MethodSpec.Builder beforeEach = MethodSpec.methodBuilder("afterAll")
-                .addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-                .returns(TypeName.VOID)
-                .addAnnotation(AfterAll.class);
-
-        if (jUnitGeneratorSettings.isDropTablesAfterRunning()) {
-            beforeEach.addStatement("dropTables()");
-        }
-        return beforeEach.build();
-    }
-
-    /**
-     * Generates a unit test method specification for a path in the test data.
-     * General structure of the test:
-     * <p>
-     * - Arrange: insert data into database.
-     * - Act: Execute select statement on database.
-     * - Assert: Assert that a result is given back from the database.
-     * <p>
-     * In order to execute SQL queries, this method will reference the
-     * generated runSQL method.
-     *
-     * @param path          The path to generate a test for.
-     * @param vendorOptions The vendor options to use.
-     * @return A method specification for a single unit test.
-     */
-    private MethodSpec generatePathTest(Path path, VendorOptions vendorOptions) {
-        // Method signature
-        MethodSpec.Builder pTestBuilder = MethodSpec.methodBuilder(
-                String.format("generatedTest%d", path.getPathNumber()));
-        pTestBuilder.addModifiers(Modifier.PUBLIC)
-                .returns(TypeName.VOID)
-                .addAnnotation(Test.class);
-
-        // Arrange part
-        pTestBuilder.addComment("Arrange: set up the fixture data");
-        InsertionBuilder insertionBuilder = new InsertionBuilder(vendorOptions);
-        for (String s : insertionBuilder.buildQueries(path)) {
-            pTestBuilder.addStatement("runSQL($S)", s);
-        }
-
-        // Act
-        pTestBuilder.addComment("Act: run a selection query on the database");
-        SelectionBuilder selectionBuilder = new SelectionBuilder(vendorOptions);
-        String select = selectionBuilder.buildQueries(path).get(0);
-        pTestBuilder.addStatement("$T result = runSQL($S)", boolean.class, select);
-
-        // Assert
-        pTestBuilder.addComment("Assert: verify that at least one resulting row was returned");
-        pTestBuilder.addStatement("$T.assertTrue(result)", Assertions.class);
-        return pTestBuilder.build();
+    public JUnit5TestGenerator(JUnitGeneratorSettings jUnitGeneratorSettings) {
+        super(
+                jUnitGeneratorSettings,
+                false,
+                Test.class,
+                BeforeAll.class,
+                AfterAll.class,
+                BeforeEach.class,
+                AfterEach.class,
+                Assertions.class
+        );
     }
 }
