+++ brew/src/test/java/nl/tudelft/serg/evosql/brew/generator/junit/JUnit5TestGeneratorTest.java
@@ -1,0 +1,54 @@
+package nl.tudelft.serg.evosql.brew.generator.junit;
+
+import nl.tudelft.serg.evosql.brew.DataGenerator;
+import nl.tudelft.serg.evosql.brew.data.Result;
+import nl.tudelft.serg.evosql.brew.db.ConnectionData;
+import nl.tudelft.serg.evosql.brew.sql.vendor.MySQLOptions;
+import org.junit.jupiter.api.Test;
+
+public class JUnit5TestGeneratorTest {
+    private Result resultSmall = DataGenerator.makeResult1();
+    private Result resultMedium = DataGenerator.makeResult2();
+
+    @Test
+    public void testMethodgenerationSmall() {
+        ConnectionData connectionData = new ConnectionData(
+                "connection string",
+                "db",
+                "user",
+                "password"
+        );
+        JUnitGeneratorSettings JUnitGeneratorSettings = new JUnitGeneratorSettings(
+                connectionData,
+                "brew.test.package",
+                true,
+                true,
+                true,
+                true,
+                true
+        );
+        JUnit5TestGenerator jUnit4TestGenerator = new JUnit5TestGenerator(JUnitGeneratorSettings);
+        System.out.println(jUnit4TestGenerator.generate(resultSmall, new MySQLOptions()));
+    }
+
+    @Test
+    public void testMethodgenerationMedium() {
+        ConnectionData connectionData = new ConnectionData(
+                "connection string",
+                "db",
+                "user",
+                "password"
+        );
+        JUnitGeneratorSettings JUnitGeneratorSettings = new JUnitGeneratorSettings(
+                connectionData,
+                "brew.test.package",
+                true,
+                true,
+                true,
+                true,
+                true
+        );
+        JUnit5TestGenerator jUnit4TestGenerator = new JUnit5TestGenerator(JUnitGeneratorSettings);
+        System.out.println(jUnit4TestGenerator.generate(resultMedium, new MySQLOptions()));
+    }
+}
